GROK-15756: Diff Studio | Solver: Refactoring
Reorganized code due to add the new method
